+++ app/routes.php
@@ -6,12 +6,14 @@ use Slim\App;
 return function (App $app) {
     $app->get('/', 'toDoListPageController');
     $app->get('/login', 'loginPageController');
+    $app->post('/login', 'authenticationController');
 
-    $app->post('/editAll', 'editAllTasksController');
     $app->post('/add', 'insertNewTaskController');
     $app->post('/complete', 'markTasksCompleteController');
     $app->post('/incomplete', 'markTasksIncompleteController');
     $app->post('/archive', 'markTasksArchivedController');
     $app->post('/recover', 'markTasksNotArchivedController');
     $app->post('/deletePermanently', 'deleteTasksPermanentlyController');
+
+	$app->post('/editAll', 'editAllTasksController');
 };
