+++ src/Controllers/DatabaseControllers/EditAllTasksController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controllers;
+namespace App\Controllers\DatabaseControllers;
 use Psr\Container\ContainerInterface;
 
 class EditAllTasksController
