+++ src/Controllers/PageControllers/LoginPageController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Controllers;
+namespace App\Controllers\PageControllers;
 use Psr\Container\ContainerInterface;
 
 class LoginPageController
